refactor(scheduled): 提取版本化存储层到 stores 子模块
