+++ pinvou3-app/src-tauri/src/features/scheduled/tasks.rs
@@ -1,5 +1,5 @@
 use std::collections::{HashMap, HashSet};
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 use std::sync::{Arc, Weak};
 use std::time::Duration;
 
@@ -12,7 +12,7 @@ use deepseek_tui::automation_manager::{
     UpdateAutomationRequest,
 };
 use deepseek_tui::task_manager::{SharedTaskManager, TaskManager, TaskManagerConfig, TaskStatus};
-use parking_lot::{Mutex as ParkingMutex, RwLock};
+use parking_lot::Mutex as ParkingMutex;
 use tokio_util::sync::CancellationToken;
 
 use crate::features::assistant::engine_pool::EnginePool;
@@ -29,530 +29,9 @@ const SCHEDULED_MODEL_BINDING_SCHEMA_VERSION: u32 = 1;
 const SCHEDULED_TASK_UI_METADATA_SCHEMA_VERSION: u32 = 1;
 const SCHEDULED_EXECUTION_MODE: &str = "yolo";
 
-fn scheduled_run_read_state_schema_version() -> u32 {
-    SCHEDULED_RUN_READ_STATE_SCHEMA_VERSION
-}
-
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-struct ScheduledRunReadRegistry {
-    #[serde(default = "scheduled_run_read_state_schema_version")]
-    schema_version: u32,
-    #[serde(default)]
-    viewed_runs: HashMap<String, HashSet<String>>,
-}
-
-impl Default for ScheduledRunReadRegistry {
-    fn default() -> Self {
-        Self {
-            schema_version: SCHEDULED_RUN_READ_STATE_SCHEMA_VERSION,
-            viewed_runs: HashMap::new(),
-        }
-    }
-}
-
-fn scheduled_model_binding_schema_version() -> u32 {
-    SCHEDULED_MODEL_BINDING_SCHEMA_VERSION
-}
-
-fn scheduled_task_ui_metadata_schema_version() -> u32 {
-    SCHEDULED_TASK_UI_METADATA_SCHEMA_VERSION
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
-struct ScheduledTaskModelBinding {
-    model_id: String,
-    model: String,
-    updated_at: String,
-}
-
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-struct ScheduledTaskModelBindingRegistry {
-    #[serde(default = "scheduled_model_binding_schema_version")]
-    schema_version: u32,
-    #[serde(default)]
-    tasks: HashMap<String, ScheduledTaskModelBinding>,
-}
-
-impl Default for ScheduledTaskModelBindingRegistry {
-    fn default() -> Self {
-        Self {
-            schema_version: SCHEDULED_MODEL_BINDING_SCHEMA_VERSION,
-            tasks: HashMap::new(),
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
-struct ScheduledTaskUiMetadata {
-    #[serde(default)]
-    pinned: bool,
-    #[serde(default)]
-    pinned_at: Option<String>,
-    updated_at: String,
-}
-
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-struct ScheduledTaskUiMetadataRegistry {
-    #[serde(default = "scheduled_task_ui_metadata_schema_version")]
-    schema_version: u32,
-    #[serde(default)]
-    tasks: HashMap<String, ScheduledTaskUiMetadata>,
-}
-
-impl Default for ScheduledTaskUiMetadataRegistry {
-    fn default() -> Self {
-        Self {
-            schema_version: SCHEDULED_TASK_UI_METADATA_SCHEMA_VERSION,
-            tasks: HashMap::new(),
-        }
-    }
-}
-
-/// How [`VersionedJsonStore`] reacts to an unreadable / unsupported payload.
-enum QuarantineStrategy {
-    /// Emit a `warn!` and leave the offending file in place (UI metadata store).
-    LogInPlace,
-    /// Rename the file to `<name>.invalid-<ts>` and log (model-binding / read-state).
-    Rename,
-}
-
-/// Per-store behaviour carried by the registry type itself.
-///
-/// The three scheduled registries share an identical open/persist skeleton but
-/// differ in (a) schema version, (b) how an old version is migrated, (c) whether
-/// invalid payloads are quarantined or merely logged, and (d) the human-readable
-/// label/suffix used in diagnostics. This trait carries exactly those
-/// differences so [`VersionedJsonStore<T>`] can stay generic without assuming
-/// the three stores are textually identical.
-trait VersionedRegistry:
-    Default + serde::Serialize + serde::de::DeserializeOwned + Clone + Send + Sync + 'static
-{
-    /// Schema version persisted in (and supported by) this registry.
-    const SUPPORTED_VERSION: u32;
-    /// Read back the schema version of a deserialised instance.
-    fn schema_version(&self) -> u32;
-    /// Migrate an older-version instance up to [`SUPPORTED_VERSION`].
-    ///
-    /// Infallible: every concrete store always produces a replacement (the
-    /// read-state store deliberately resets to default, dropping viewed runs).
-    fn migrate(self) -> Self;
-    /// Quarantine policy for newer-than-supported / invalid-JSON payloads.
-    const QUARANTINE: QuarantineStrategy;
-    /// Human-readable label used in log/error messages for this store.
-    const LABEL: &'static str;
-    /// Fallback file-name stem when the on-disk path has no file component.
-    const QUARANTINE_FALLBACK_NAME: &'static str;
-    /// Store-specific suffix appended to quarantine / read-failure warnings.
-    const WARN_SUFFIX: &'static str;
-}
-
-/// Versioned-JSON registry store with schema migration, quarantine, and atomic
-/// writes. Collapses the three previously hand-rolled stores into one generic
-/// core; per-store differences live on [`VersionedRegistry`].
-#[derive(Clone)]
-struct VersionedJsonStore<T: VersionedRegistry> {
-    path: Arc<PathBuf>,
-    registry: Arc<RwLock<T>>,
-}
-
-impl<T: VersionedRegistry> VersionedJsonStore<T> {
-    fn open(path: PathBuf) -> Result<Self> {
-        let mut migrated = false;
-        let registry = match std::fs::read_to_string(&path) {
-            Ok(raw) => match serde_json::from_str::<T>(&raw) {
-                Ok(registry) if registry.schema_version() == T::SUPPORTED_VERSION => registry,
-                Ok(registry) if registry.schema_version() < T::SUPPORTED_VERSION => {
-                    migrated = true;
-                    registry.migrate()
-                }
-                Ok(registry) => {
-                    Self::handle_invalid(
-                        &path,
-                        &format!(
-                            "schema v{} is newer than supported v{}",
-                            registry.schema_version(),
-                            T::SUPPORTED_VERSION
-                        ),
-                    );
-                    T::default()
-                }
-                Err(error) => {
-                    Self::handle_invalid(&path, &format!("invalid JSON: {error}"));
-                    T::default()
-                }
-            },
-            Err(error) if error.kind() == std::io::ErrorKind::NotFound => T::default(),
-            Err(error) => {
-                log::warn!(
-                    "Unable to read {} {}: {error}{}",
-                    T::LABEL,
-                    path.display(),
-                    T::WARN_SUFFIX
-                );
-                T::default()
-            }
-        };
-        let store = Self {
-            path: Arc::new(path),
-            registry: Arc::new(RwLock::new(registry)),
-        };
-        if migrated {
-            if let Err(error) = store.persist(&store.registry.read()) {
-                log::warn!(
-                    "Unable to persist migrated {} {}: {error:#}",
-                    T::LABEL,
-                    store.path.display()
-                );
-            }
-        }
-        Ok(store)
-    }
-
-    fn persist(&self, registry: &T) -> Result<()> {
-        if let Some(parent) = self.path.parent() {
-            std::fs::create_dir_all(parent)
-                .with_context(|| format!("create {} dir {}", T::LABEL, parent.display()))?;
-        }
-        let payload = serde_json::to_vec_pretty(registry)
-            .with_context(|| format!("serialize {}", T::LABEL))?;
-        deepseek_tui::utils::write_atomic(self.path.as_ref(), &payload)
-            .with_context(|| format!("write {} {}", T::LABEL, self.path.display()))
-    }
-
-    /// Apply this store's quarantine policy to an invalid payload at `path`.
-    fn handle_invalid(path: &Path, reason: &str) {
-        match T::QUARANTINE {
-            QuarantineStrategy::LogInPlace => {
-                log::warn!(
-                    "Ignoring invalid {} {} ({reason})",
-                    T::LABEL,
-                    path.display()
-                );
-            }
-            QuarantineStrategy::Rename => {
-                let timestamp = std::time::SystemTime::now()
-                    .duration_since(std::time::UNIX_EPOCH)
-                    .unwrap_or_default()
-                    .as_nanos();
-                let file_name = path
-                    .file_name()
-                    .and_then(|name| name.to_str())
-                    .unwrap_or(T::QUARANTINE_FALLBACK_NAME);
-                let quarantine_path =
-                    path.with_file_name(format!("{file_name}.invalid-{timestamp}"));
-                match std::fs::rename(path, &quarantine_path) {
-                    Ok(()) => log::warn!(
-                        "Quarantined {} {} to {} ({reason}){}",
-                        T::LABEL,
-                        path.display(),
-                        quarantine_path.display(),
-                        T::WARN_SUFFIX
-                    ),
-                    Err(error) => log::warn!(
-                        "Invalid {} {} ({reason}) could not be quarantined: {error}{}",
-                        T::LABEL,
-                        path.display(),
-                        T::WARN_SUFFIX
-                    ),
-                }
-            }
-        }
-    }
-}
-
-impl VersionedRegistry for ScheduledTaskUiMetadataRegistry {
-    const SUPPORTED_VERSION: u32 = SCHEDULED_TASK_UI_METADATA_SCHEMA_VERSION;
-    const QUARANTINE: QuarantineStrategy = QuarantineStrategy::LogInPlace;
-    const LABEL: &'static str = "scheduled task UI metadata";
-    const QUARANTINE_FALLBACK_NAME: &'static str = "scheduled-task-ui-metadata.json";
-    const WARN_SUFFIX: &'static str = "";
-
-    fn schema_version(&self) -> u32 {
-        self.schema_version
-    }
-
-    fn migrate(self) -> Self {
-        Self {
-            schema_version: SCHEDULED_TASK_UI_METADATA_SCHEMA_VERSION,
-            tasks: self.tasks,
-        }
-    }
-}
-
-impl VersionedRegistry for ScheduledTaskModelBindingRegistry {
-    const SUPPORTED_VERSION: u32 = SCHEDULED_MODEL_BINDING_SCHEMA_VERSION;
-    const QUARANTINE: QuarantineStrategy = QuarantineStrategy::Rename;
-    const LABEL: &'static str = "scheduled model binding state";
-    const QUARANTINE_FALLBACK_NAME: &'static str = "model-bindings.json";
-    const WARN_SUFFIX: &'static str = "; scheduled tasks will fall back to wire model names";
-
-    fn schema_version(&self) -> u32 {
-        self.schema_version
-    }
-
-    fn migrate(self) -> Self {
-        Self {
-            schema_version: SCHEDULED_MODEL_BINDING_SCHEMA_VERSION,
-            tasks: self.tasks,
-        }
-    }
-}
-
-impl VersionedRegistry for ScheduledRunReadRegistry {
-    const SUPPORTED_VERSION: u32 = SCHEDULED_RUN_READ_STATE_SCHEMA_VERSION;
-    const QUARANTINE: QuarantineStrategy = QuarantineStrategy::Rename;
-    const LABEL: &'static str = "scheduled run read state";
-    const QUARANTINE_FALLBACK_NAME: &'static str = "scheduled-run-read-state.json";
-    const WARN_SUFFIX: &'static str = "; treating all runs as unread";
-
-    fn schema_version(&self) -> u32 {
-        self.schema_version
-    }
-
-    /// Older read-state schemas are dropped: viewed-run tracking resets on
-    /// upgrade, matching the legacy hand-rolled store.
-    fn migrate(self) -> Self {
-        Self::default()
-    }
-}
-
-type ScheduledTaskUiMetadataStore = VersionedJsonStore<ScheduledTaskUiMetadataRegistry>;
-
-impl VersionedJsonStore<ScheduledTaskUiMetadataRegistry> {
-    fn metadata_for(&self, automation_id: &str) -> (bool, Option<String>) {
-        self.registry
-            .read()
-            .tasks
-            .get(automation_id)
-            .filter(|metadata| metadata.pinned)
-            .map(|metadata| (true, metadata.pinned_at.clone()))
-            .unwrap_or((false, None))
-    }
-
-    fn set_pinned(&self, automation_id: &str, pinned: bool) -> Result<()> {
-        if automation_id.trim().is_empty() {
-            bail!("scheduled automation id cannot be empty");
-        }
-        let now = chrono::Utc::now().to_rfc3339();
-        let mut registry = self.registry.write();
-        let previous = registry.tasks.get(automation_id).cloned();
-        if pinned {
-            registry.tasks.insert(
-                automation_id.to_string(),
-                ScheduledTaskUiMetadata {
-                    pinned: true,
-                    pinned_at: Some(now.clone()),
-                    updated_at: now,
-                },
-            );
-        } else {
-            registry.tasks.remove(automation_id);
-        }
-        if let Err(error) = self.persist(&registry) {
-            match previous {
-                Some(metadata) => {
-                    registry.tasks.insert(automation_id.to_string(), metadata);
-                }
-                None => {
-                    registry.tasks.remove(automation_id);
-                }
-            }
-            return Err(error);
-        }
-        Ok(())
-    }
-
-    fn remove(&self, automation_id: &str) -> Result<()> {
-        let mut registry = self.registry.write();
-        let Some(previous) = registry.tasks.remove(automation_id) else {
-            return Ok(());
-        };
-        if let Err(error) = self.persist(&registry) {
-            registry.tasks.insert(automation_id.to_string(), previous);
-            return Err(error);
-        }
-        Ok(())
-    }
-
-    fn compact(&self, automation_ids: &HashSet<String>) -> Result<()> {
-        let mut registry = self.registry.write();
-        let before = registry.tasks.clone();
-        registry.tasks.retain(|id, _| automation_ids.contains(id));
-        if registry.tasks == before {
-            return Ok(());
-        }
-        if let Err(error) = self.persist(&registry) {
-            registry.tasks = before;
-            return Err(error);
-        }
-        Ok(())
-    }
-}
-
-type ScheduledTaskModelBindingStore = VersionedJsonStore<ScheduledTaskModelBindingRegistry>;
-
-impl VersionedJsonStore<ScheduledTaskModelBindingRegistry> {
-    fn model_id_for(&self, automation_id: &str, model: &str) -> Option<String> {
-        self.registry
-            .read()
-            .tasks
-            .get(automation_id)
-            .filter(|binding| binding.model == model)
-            .map(|binding| binding.model_id.clone())
-    }
-
-    fn set(
-        &self,
-        automation_id: &str,
-        model_id: Option<String>,
-        model: Option<String>,
-    ) -> Result<()> {
-        if automation_id.trim().is_empty() {
-            bail!("scheduled automation id cannot be empty");
-        }
-        let model_id = model_id
-            .map(|value| value.trim().to_string())
-            .filter(|value| !value.is_empty());
-        let model = model
-            .map(|value| value.trim().to_string())
-            .filter(|value| !value.is_empty());
-        let mut registry = self.registry.write();
-        let previous = registry.tasks.get(automation_id).cloned();
-        match (model_id, model) {
-            (Some(model_id), Some(model)) => {
-                registry.tasks.insert(
-                    automation_id.to_string(),
-                    ScheduledTaskModelBinding {
-                        model_id,
-                        model,
-                        updated_at: chrono::Utc::now().to_rfc3339(),
-                    },
-                );
-            }
-            _ => {
-                registry.tasks.remove(automation_id);
-            }
-        }
-        if let Err(error) = self.persist(&registry) {
-            match previous {
-                Some(binding) => {
-                    registry.tasks.insert(automation_id.to_string(), binding);
-                }
-                None => {
-                    registry.tasks.remove(automation_id);
-                }
-            }
-            return Err(error);
-        }
-        Ok(())
-    }
-
-    fn remove(&self, automation_id: &str) -> Result<()> {
-        let mut registry = self.registry.write();
-        let Some(previous) = registry.tasks.remove(automation_id) else {
-            return Ok(());
-        };
-        if let Err(error) = self.persist(&registry) {
-            registry.tasks.insert(automation_id.to_string(), previous);
-            return Err(error);
-        }
-        Ok(())
-    }
-
-    fn compact(&self, automation_ids: &HashSet<String>) -> Result<()> {
-        let mut registry = self.registry.write();
-        let before = registry.tasks.clone();
-        registry.tasks.retain(|id, _| automation_ids.contains(id));
-        if registry.tasks == before {
-            return Ok(());
-        }
-        if let Err(error) = self.persist(&registry) {
-            registry.tasks = before;
-            return Err(error);
-        }
-        Ok(())
-    }
-}
-
-type ScheduledRunReadStore = VersionedJsonStore<ScheduledRunReadRegistry>;
-
-impl VersionedJsonStore<ScheduledRunReadRegistry> {
-    fn is_viewed(&self, automation_id: &str, run_id: &str) -> bool {
-        self.registry
-            .read()
-            .viewed_runs
-            .get(automation_id)
-            .is_some_and(|runs| runs.contains(run_id))
-    }
-
-    fn mark_viewed(&self, automation_id: &str, run_id: &str) -> Result<()> {
-        if automation_id.trim().is_empty() || run_id.trim().is_empty() {
-            bail!("scheduled automation and run ids cannot be empty");
-        }
-        let mut registry = self.registry.write();
-        let inserted = registry
-            .viewed_runs
-            .entry(automation_id.to_string())
-            .or_default()
-            .insert(run_id.to_string());
-        if !inserted {
-            return Ok(());
-        }
-        if let Err(error) = self.persist(&registry) {
-            if let Some(runs) = registry.viewed_runs.get_mut(automation_id) {
-                runs.remove(run_id);
-                if runs.is_empty() {
-                    registry.viewed_runs.remove(automation_id);
-                }
-            }
-            return Err(error);
-        }
-        Ok(())
-    }
-
-    fn remove_automation(&self, automation_id: &str) -> Result<()> {
-        let mut registry = self.registry.write();
-        let Some(removed) = registry.viewed_runs.remove(automation_id) else {
-            return Ok(());
-        };
-        if let Err(error) = self.persist(&registry) {
-            registry
-                .viewed_runs
-                .insert(automation_id.to_string(), removed);
-            return Err(error);
-        }
-        Ok(())
-    }
-
-    fn compact(&self, automation_id: &str, current_run_ids: &HashSet<String>) -> Result<()> {
-        let mut registry = self.registry.write();
-        let Some(existing) = registry.viewed_runs.get(automation_id).cloned() else {
-            return Ok(());
-        };
-        let retained = existing
-            .iter()
-            .filter(|run_id| current_run_ids.contains(*run_id))
-            .cloned()
-            .collect::<HashSet<_>>();
-        if retained == existing {
-            return Ok(());
-        }
-        if retained.is_empty() {
-            registry.viewed_runs.remove(automation_id);
-        } else {
-            registry
-                .viewed_runs
-                .insert(automation_id.to_string(), retained);
-        }
-        if let Err(error) = self.persist(&registry) {
-            registry
-                .viewed_runs
-                .insert(automation_id.to_string(), existing);
-            return Err(error);
-        }
-        Ok(())
-    }
-}
+#[path = "stores.rs"]
+mod stores;
+use stores::*;
 
 /// 删除一次定时运行对话的那一步。抽成 trait 只为可注入：EnginePool 需要活的
 /// WebView AppHandle，无法在单元测试里构造，而删除级联的正确性必须被测到。
@@ -2127,6 +1606,7 @@ pub fn scheduled_task_chat_prompt() -> Result<String, String> {
 #[allow(clippy::await_holding_lock)]
 mod tests {
     use super::*;
+    use parking_lot::RwLock;
 
     impl ScheduledTaskState {
         async fn create_for_test(
